Add luxon library, date and refactor
