+++ index.html
@@ -19,7 +19,7 @@
             </ul>
         </nav>
     </header>
-    <p class="time"></p>
+    <p id="time" class="time"></p>
     <main>
         <section id="list">
             <h1 class="heading">Awesome Books</h1>
